requirements.txt
# requirements.txt

certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.3
narwhals==1.47.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pillow==10.4.0
plotly==6.2.0
pluggy==1.6.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
schedule==1.2.2
six==1.17.0
tkhtmlview==0.3.1
tzdata==2025.2
urllib3==2.5.0
plotly>=5.0.0
pandas
requests
Pillow
tkinterweb
tkhtmlview

# ─── New dependencies ───────────────────────────────────────────────────────────
scikit-learn>=0.24.0
joblib>=1.1.0
mplcursors>=0.5.0
